+++ src/fmat33.hpp
@@ -4,14 +4,14 @@
 
 struct fmat33 final
 {
-	fmat33(const float value);
+	fmat33(const double value);
 
 	const fmat33 operator*(const fmat33 & other) const;
 	const fvec2 operator*(const fvec2 & other) const;
 
 	union
 	{
-		float m[3][3];
-		float f[9];
+		double m[3][3];
+		double f[9];
 	};
 };
