+++ src/quad.hpp
@@ -16,13 +16,13 @@ public:
 	Quad(const shared_ptr<VertexBuffer> vertexBuffer, const shared_ptr<TextureBuffer> textureBuffer, int depth);
 
 	void update();
-	void setOpacity(const float value);
+	void setOpacity(const double value);
 	void setColor(const fvec3 & value);
 	void setPosition(const fvec2 & value);
-	void setRotation(const float value);
+	void setRotation(const double value);
 	void setSize(const fvec2 & value);
 	void setVisible(const bool value);
-	void setLightness(const float value);
+	void setLightness(const double value);
 	void setUvMultiplier(const fvec2 & value);
 	void setUvOffset(const fvec2 & value);
 
@@ -38,9 +38,9 @@ public:
 	const fvec2 & getUvMultiplier() const;
 	const fvec2 & getUvOffset() const;
 
-	const float getOpacity() const;
-	const float getRotation() const;
-	const float getLightness() const;
+	const double getOpacity() const;
+	const double getRotation() const;
+	const double getLightness() const;
 
 	const int getDepth() const;
 
@@ -61,9 +61,9 @@ private:
 	fvec2 _uvMultiplier = fvec2(1.0);
 	fvec2 _uvOffset = fvec2(0.0);
 
-	float _rotation = 0.0;
-	float _opacity = 1.0;
-	float _lightness = 1.0;
+	double _rotation = 0.0;
+	double _opacity = 1.0;
+	double _lightness = 1.0;
 
 	bool _isVisible = true;
 };
